Move pemrission tests
